+++ e2e/specs/administration/user.spec.ts
@@ -32,7 +32,7 @@ describe('Full Admin Administartion User Flow', () => {
 
     beforeAll(async () => {
         await logIn.logInAs(users.admin.user_name, users.admin.password);
-        return (await projectList.menuBar.user()).administration();
+        return projectList.menuBar.administration();
     });
 
     afterAll(async () => {
@@ -100,7 +100,7 @@ describe('Full Admin Administartion User Flow', () => {
             await expect(projectList.isOpened()).toBe(true, 'New user is not able to log in');
             await projectList.menuBar.clickLogOut();
             await logIn.logInAs(users.admin.user_name, users.admin.password);
-            return (await projectList.menuBar.user()).administration();
+            return projectList.menuBar.administration();
         });
     });
 
@@ -165,7 +165,7 @@ describe('Full Admin Administartion User Flow', () => {
             await expect(projectList.isOpened()).toBe(true, 'New user is not able to log in');
             await projectList.menuBar.clickLogOut();
             await logIn.logInAs(users.admin.user_name, users.admin.password);
-            return (await projectList.menuBar.user()).administration();
+            return projectList.menuBar.administration();
         });
     });